setPrice was long, changed it to double

+++ demo - 1/demo/src/main/java/com/example/demo/content/Book.java
@@ -64,7 +64,7 @@ public long getId() {
 					    return price;
 					  }
 
-					  public void setPrice(long price) {
+					  public void setPrice(double price) {
 					    this.price = price;
 					  }
